+++ app/src/main/java/sg/go/user/Models/TypeCarRequest.java
@@ -3,13 +3,13 @@ package sg.go.user.Models;
 public class TypeCarRequest {
     String Name_Type_Car_Request;
     String Imga_Type_Car_Request;
-    int Name_Type_service_fee;
+    double Name_Type_service_fee;
 
-    public int getName_Type_service_fee() {
+    public double getName_Type_service_fee() {
         return Name_Type_service_fee;
     }
 
-    public void setName_Type_service_fee(int name_Type_service_fee) {
+    public void setName_Type_service_fee(double name_Type_service_fee) {
         Name_Type_service_fee = name_Type_service_fee;
     }
 
@@ -24,7 +24,7 @@ public class TypeCarRequest {
         Imga_Type_Car_Request_Int = imga_Type_Car_Request_Int;
     }
 
-    public TypeCarRequest(String name_Type_Car_Request, String imga_Type_Car_Request, int name_Type_service_fee) {
+    public TypeCarRequest(String name_Type_Car_Request, String imga_Type_Car_Request, double name_Type_service_fee) {
         Name_Type_Car_Request = name_Type_Car_Request;
         Imga_Type_Car_Request = imga_Type_Car_Request;
         Name_Type_service_fee = name_Type_service_fee;
